+++ R/mutate.R
@@ -73,7 +73,7 @@ mutate.DataFrame <- function(.data, ...) {
   is_grouped <- !is.null(grps)
   to_drop <- list()
 
-  polars_exprs <- translate_dots(.data = .data, ..., env = rlang::caller_env())
+  polars_exprs <- translate_dots(.data = .data, ..., env = rlang::current_env())
 
   for (i in seq_along(polars_exprs)) {
     sub <- polars_exprs[[i]]